+++ controllers/kafkaconnect_controller.go
@@ -6,7 +6,7 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/aiven/aiven-go-client"
+	"github.com/aiven/aiven-go-client/v2"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	ctrl "sigs.k8s.io/controller-runtime"
@@ -63,7 +63,7 @@ func (a *kafkaConnectAdapter) getUserConfig() any {
 	return &a.Spec.UserConfig
 }
 
-func (a *kafkaConnectAdapter) newSecret(_ *aiven.Service) (*corev1.Secret, error) {
+func (a *kafkaConnectAdapter) newSecret(ctx context.Context, s *aiven.Service) (*corev1.Secret, error) {
 	return nil, nil
 }