+++ controllers/mysql_controller.go
@@ -6,7 +6,7 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/aiven/aiven-go-client"
+	"github.com/aiven/aiven-go-client/v2"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	ctrl "sigs.k8s.io/controller-runtime"
@@ -65,7 +65,7 @@ func (a *mySQLAdapter) getUserConfig() any {
 	return &a.Spec.UserConfig
 }
 
-func (a *mySQLAdapter) newSecret(s *aiven.Service) (*corev1.Secret, error) {
+func (a *mySQLAdapter) newSecret(ctx context.Context, s *aiven.Service) (*corev1.Secret, error) {
 	stringData := map[string]string{
 		"HOST":        s.URIParams["host"],
 		"PORT":        s.URIParams["port"],