+++ tests/connectionpool_test.go
@@ -1,6 +1,7 @@
 package tests
 
 import (
+	"context"
 	"fmt"
 	"testing"
 
@@ -78,6 +79,7 @@ func TestConnectionPool(t *testing.T) {
 	defer recoverPanic(t)
 
 	// GIVEN
+	ctx := context.Background()
 	pgName := randName("connection-pool")
 	dbName := randName("connection-pool")
 	userName := randName("connection-pool")
@@ -107,7 +109,7 @@ func TestConnectionPool(t *testing.T) {
 
 	// THEN
 	// Validates PostgreSQL
-	pgAvn, err := avnClient.Services.Get(testProject, pgName)
+	pgAvn, err := avnClient.Services.Get(ctx, testProject, pgName)
 	require.NoError(t, err)
 	assert.Equal(t, pgAvn.Name, pg.GetName())
 	assert.Equal(t, "RUNNING", pg.Status.State)
@@ -116,20 +118,20 @@ func TestConnectionPool(t *testing.T) {
 	assert.Equal(t, pgAvn.CloudName, pg.Spec.CloudName)
 
 	// Validates Database
-	dbAvn, err := avnClient.Databases.Get(testProject, pgName, dbName)
+	dbAvn, err := avnClient.Databases.Get(ctx, testProject, pgName, dbName)
 	require.NoError(t, err)
 	assert.Equal(t, dbName, db.GetName())
 	assert.Equal(t, dbAvn.DatabaseName, db.GetName())
 
 	// Validates ServiceUser
-	userAvn, err := avnClient.ServiceUsers.Get(testProject, pgName, userName)
+	userAvn, err := avnClient.ServiceUsers.Get(ctx, testProject, pgName, userName)
 	require.NoError(t, err)
 	assert.Equal(t, userName, user.GetName())
 	assert.Equal(t, userName, userAvn.Username)
 	assert.Equal(t, pgName, user.Spec.ServiceName)
 
 	// Validates ConnectionPool
-	poolAvn, err := avnClient.ConnectionPools.Get(testProject, pgName, poolName)
+	poolAvn, err := avnClient.ConnectionPools.Get(ctx, testProject, pgName, poolName)
 	require.NoError(t, err)
 	assert.Equal(t, pgName, pool.Spec.ServiceName)
 	assert.Equal(t, poolName, pool.GetName())
@@ -168,7 +170,7 @@ func TestConnectionPool(t *testing.T) {
 	// if service is deleted, pool is destroyed in Aiven. No service — no pool. No pool — no pool.
 	// And we make sure that controller can delete db itself
 	assert.NoError(t, s.Delete(pool, func() error {
-		_, err = avnClient.ConnectionPools.Get(testProject, pgName, poolName)
+		_, err = avnClient.ConnectionPools.Get(ctx, testProject, pgName, poolName)
 		return err
 	}))
 }